2k-301 Add computation expression with error handling

+++ tests/HW5.Tests/AssertHelper.fs
@@ -1,4 +1,4 @@
 module HW5.Tests.AssertHelper
 open Xunit
-let ass (expected: int) (actual: int) =
+let ass (expected: decimal) (actual: decimal) =
     Assert.Equal (expected, actual)
